refactor(types): :label: add `ApiQueryStringSchema`, update type imports
